Move main method generation to new class
More refactoring by moving the main method generation code to a separate class.
